For #660: package io.shardingjdbc.test.sql.jaxb => io.shardingjdbc.test.sql
