+++ sharding-core/src/test/java/io/shardingjdbc/core/parsing/UnsupportedSQLParsingEngineTest.java
@@ -22,8 +22,8 @@ import com.google.common.collect.Sets;
 import io.shardingjdbc.core.api.fixture.ShardingRuleMockBuilder;
 import io.shardingjdbc.core.constant.DatabaseType;
 import io.shardingjdbc.core.parsing.parser.exception.SQLParsingUnsupportedException;
-import io.shardingjdbc.test.sql.jaxb.SQLCase;
-import io.shardingjdbc.test.sql.jaxb.SQLCasesLoader;
+import io.shardingjdbc.test.sql.SQLCase;
+import io.shardingjdbc.test.sql.SQLCasesLoader;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
