+++ sharding-jdbc-core/src/test/java/io/shardingjdbc/core/integrate/jaxb/helper/SQLAssertJAXBHelper.java
@@ -21,7 +21,7 @@ import io.shardingjdbc.core.constant.DatabaseType;
 import io.shardingjdbc.core.constant.SQLType;
 import io.shardingjdbc.core.integrate.jaxb.SQLAssert;
 import io.shardingjdbc.core.integrate.jaxb.SQLAsserts;
-import io.shardingjdbc.test.sql.jaxb.SQLCasesLoader;
+import io.shardingjdbc.test.sql.SQLCasesLoader;
 
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
