+++ sharding-sql-test/src/main/java/io/shardingjdbc/test/sql/SQLCases.java
@@ -15,31 +15,24 @@
  * </p>
  */
 
-package io.shardingjdbc.test.sql.jaxb;
+package io.shardingjdbc.test.sql;
 
 import lombok.Getter;
-import lombok.Setter;
 
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlRootElement;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
- * SQL test case for xml tag.
- *
- * @author zhangliang
+ * SQL test cases for xml root tag.
+ * 
+ * @author zhangliang 
  */
-@XmlAccessorType(XmlAccessType.FIELD)
+@XmlRootElement(name = "sql-cases")
 @Getter
-@Setter
-public final class SQLCase {
+public final class SQLCases {
     
-    @XmlAttribute
-    private String id;
-    
-    @XmlAttribute
-    private String value;
-    
-    @XmlAttribute(name = "db-types")
-    private String databaseTypes;
+    @XmlElement(name = "sql-case")
+    private List<SQLCase> sqlCases = new ArrayList<>();
 }
